[ADD]: new middleware
-- In this commit --

A new middleware has been added to the server side. This middleware is a rate limiter that will limit the number of requests that can be made to the server in a certain amount of time. This middleware is used to prevent abuse of the server and to prevent the server from being overwhelmed by too many requests.